requirements.txt
Django==2.0.1
pytz==2017.3
djangorestframework==3.7.7
psycopg2==2.7.3.2
requests==2.18.4
jira
atlassian-python-api
django-material
